Added all preference tab items to UI, not all actually do something but they're there. Stubbed Language a little further still not sure how that'll end up working. Fixed a lot if not all the random binding bugs that started after upgrading to preview5, stumbled onto fix which was adding empty DataContext assignment which was a weird wpf thing that's weird in the opposite way w/ avalonia, also converted that lost focus behavior to extension (which probably was erroring from the dataContext bug), streamline styles for parameters and consolidated style defs into the listboxview now so not in that theme folder. Added button parameter which I didn't want to do since it doesn't hold value but is needed for preferences so may have to figure something about that cause it shouldn't be front facing for the plugins, won't do anything. Lots of little bugs though, saving prefs and glitchy focus things but cleared out majority. Also had to add another condition for highlighting to ignore non-string match criterias where date-time have match values for day counts and they would end up getting highlighted, really works pretty well though.

+++ MonkeyPaste.Avalonia/ViewModels/Reusable/Parameters/MpAvSliderParameterViewModel.cs
@@ -8,13 +8,8 @@ namespace MonkeyPaste.Avalonia {
         #region Private Variables
         #endregion
 
-        #region Properties
+        #region Interfaces
 
-        #region View Models
-        #endregion
-
-        #region Appearance
-        #endregion
 
         #region MpISliderViewModel Implementation
 
@@ -39,9 +34,22 @@ namespace MonkeyPaste.Avalonia {
 
         #endregion
 
+        #endregion
+
+        #region Properties
+
+        #region View Models
+        #endregion
+
+        #region Appearance
+        #endregion
+
+        #region State
+        #endregion
+
         #region Model
 
-        #endregion        
+        #endregion
 
         #endregion
 
@@ -80,7 +88,7 @@ namespace MonkeyPaste.Avalonia {
             try {
                 switch (UnitType) {
                     case MpParameterValueUnitType.Integer:
-                        return int.Parse(value);
+                        return (int)double.Parse(value);
                     case MpParameterValueUnitType.Decimal:
                         return Math.Round(double.Parse(value), Precision);
                     default:
